requirements.txt
numpy>=1.19.2
opencv-python>=4.5.1
torch>=1.8.0
torchvision>=0.9.0
mtcnn>=0.1.0
facenet-pytorch>=2.5.2
scipy>=1.6.0
pymongo>=4.0.1
python-dotenv>=0.19.0
deep-sort-realtime>=1.3.0
pydantic>=2.0.0
dlib>=19.22.0
tensorflow-macos>=2.9.0; platform_system=="Darwin" and platform_machine=="arm64"
tensorflow>=2.9.0; platform_system!="Darwin" or platform_machine!="arm64"
mediapipe>=0.10.0
